# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hideout-api",
    "version": "0.1.0",
    "description": "REST contract for hideout-web. Generated from src/contracts/http; do not edit."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hideout_session",
        "description": "httpOnly session cookie, named __Host-hideout_session over HTTPS."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_FAILED",
                "description": "One of: BAD_REQUEST, INVALID_JSON, UNAUTHENTICATED, FORBIDDEN, ORIGIN_NOT_ALLOWED, UNSUPPORTED_CONTENT_TYPE, NOT_FOUND, CONFLICT, PAYLOAD_TOO_LARGE, VALIDATION_FAILED, RATE_LIMITED, INTERNAL."
              },
              "message": {
                "type": "string",
                "example": "This invite has expired."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                },
                "description": "Present on 422 validation errors."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out of this device",
        "description": "Deletes the current session and clears the cookie. Send `Content-Type: application/json` (no body needed) with credentials; the CSRF check requires it.",
        "responses": {
          "204": {
            "description": "Signed out."
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin or Content-Type check failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out of every device",
        "description": "Deletes all of the current user’s sessions and clears the cookie. Same request rules as logout.",
        "responses": {
          "204": {
            "description": "Signed out everywhere."
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin or Content-Type check failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}